Restructured the code for the Select component and renamed it
